fixed wrong import lines due to renaming
