+++ backend/src/controllers/faculty.controller.js
@@ -1,4 +1,4 @@
-import Faculties from "../models/facultymodel.js";
+import Faculties from "../models/faculty.model.js";
 import { asyncHandler } from "../utils/asyncHandler.js";
 import { ApiResponse } from "../utils/ApiResponse.js";
 import { ApiError } from "../utils/ApiError.js";
