+++ backend/src/controllers/list.controller.js
@@ -1,4 +1,4 @@
-import Lists from "../models/listmodel.js";
+import Lists from "../models/list.model.js";
 import { asyncHandler } from "../utils/asyncHandler.js";
 import { ApiResponse } from "../utils/ApiResponse.js";
 import { ApiError } from "../utils/ApiError.js";
