+++ backend/src/controllers/table.controller.js
@@ -1,4 +1,4 @@
-import Tables from "../models/tablemodel.js";
+import Tables from "../models/table.model.js";
 import { asyncHandler } from "../utils/asyncHandler.js";
 import { ApiResponse } from "../utils/ApiResponse.js";
 import { ApiError } from "../utils/ApiError.js";
