+++ backend/src/controllers/user.controller.js
@@ -1,4 +1,4 @@
-import User from "../models/usermodel.js";
+import User from "../models/user.model.js";
 import { asyncHandler } from "../utils/asyncHandler.js";
 import { ApiResponse } from "../utils/ApiResponse.js";
 import { ApiError } from "../utils/ApiError.js";
