+++ backend/src/routes/table.js
@@ -3,8 +3,9 @@ import {
     get_timetable,
     post_teachertable,
     save_timetable,
-} from "../controllers/tableController.js";
-import { verifyJWT } from "../middleware/authmiddleware.js";
+    removetable,
+} from "../controllers/table.controller.js";
+import { verifyJWT } from "../middleware/auth.middleware.js";
 
 const router = Router();
 
@@ -13,5 +14,6 @@ router.get("/get-timetable", get_timetable);
 // secured routes
 router.post("/post-teachertable", verifyJWT, post_teachertable);
 router.post("/save-timetable", verifyJWT, save_timetable);
+router.delete("/remove", verifyJWT, removetable);
 
 export default router;
